+++ NEWS.md
@@ -3,6 +3,12 @@
 * New function `pretty_print_code()` to print function objects with syntax
   highlighting at the R console.
 
+* New function `ansi_collapse()` to collapse character vectors into a single
+  string.
+
+* `ansi_strtrim()` now handles some edge cases better, when `ellipsis` has
+  length zero, and when it is wider than `width`.
+
 # cli 3.4.1
 
 * cli has better error messages now.